[Silver II] Title: A → B, Time: 140 ms, Memory: 19968 KB -BaekjoonHub

+++ 백준/Silver/16953. A → B/A → B.java
@@ -12,8 +12,8 @@ public class Main {
         BufferedReader br = new BufferedReader(new InputStreamReader(System.in));
         StringTokenizer st = new StringTokenizer(br.readLine());
 
-        long A = Integer.parseInt(st.nextToken());
-        long B = Integer.parseInt(st.nextToken());
+        long A = Long.parseLong(st.nextToken());
+        long B = Long.parseLong(st.nextToken());
 
         System.out.println(main.solution(A, B));
     }
